requirements.txt
websockets~=10.4        # Last version with good Python 3.9 support
aiohttp==3.9.1         # Stable version compatible with Python 3.9
python-dotenv==1.0.0  # Well-tested version for Python 3.9+
typing-extensions>=4.4.0  # Added for better type hint support in Python 3.9
discord-webhook==1.3.0
